// src/user/types/FormValues.ts
export type FormValuesMinuta = {
    fechaInicial: string;
    asistencia: string;
    temas: string;
    desarrollo: string;
    conclusion: string;
    pendientes: string;
    proximaFecha: string;
};


export type FormValuesInforme = {
    titulo: string;
    fechaDesde: string;
    fechaHasta: string;
    desarrollo: string;
    conclusion: string;
};